feat: useFilter Custom Hook Created
