+++ components/main/Filters.tsx
@@ -5,19 +5,25 @@ import React, { ReactNode } from 'react'
 import { STATES } from '@/lib/constants/regions'
 import { ROUTES } from '@/lib/constants/routes'
 import LucideIcon from '@/lib/icons/LucideIcon'
+import {
+  FilterType,
+  isFinishedChoices,
+  IsFinishedChoicesType,
+  stateChoices,
+  StateChoicesType,
+} from '@/lib/utils/hooks/useFilters'
 import { ReadOnly } from '@/lib/utils/typeUtils'
 
-import { FilterType, isFinishedChoices, IsFinishedChoicesType, stateChoices, StateChoicesType } from './Contents'
 import Filter from './Filter'
-
 interface FiltersProps {
   filter: FilterType
-  handleFilters: (
+  filterHandler: (
     id: 'isFinished' | 'state' | 'city' | 'all',
     type: 'change' | 'reset',
     filterValues?: FilterType['isFinished' | 'state' | 'city'],
     selectedState?: string,
   ) => void
+  movePageHandler: (pageNumber: number) => void
 }
 
 export interface FilterDisplayType {
@@ -28,7 +34,7 @@ export interface FilterDisplayType {
 }
 
 // 역할: UI 보여주고, onClick 핸들링
-const Filters = ({ filter, handleFilters }: FiltersProps): ReactNode => {
+const Filters = ({ filter, filterHandler, movePageHandler }: FiltersProps): ReactNode => {
   const pathname = usePathname()
   /**
    * URL에 따른 필터 버튼
@@ -88,7 +94,9 @@ const Filters = ({ filter, handleFilters }: FiltersProps): ReactNode => {
 
       // 첫 번째 선택된 도시가 있는 경우에만 placeHolder 생성
       const placeHolder = firstSelectedCity
-        ? `${firstSelectedState} ${firstSelectedCity} 외 ${totalSelectedCount - 1}개`
+        ? totalSelectedCount === 1
+          ? `${firstSelectedState} ${firstSelectedCity}`
+          : `${firstSelectedState} ${firstSelectedCity} 외 ${totalSelectedCount - 1}개`
         : '지역'
 
       return {
@@ -120,11 +128,12 @@ const Filters = ({ filter, handleFilters }: FiltersProps): ReactNode => {
           filter={FILTER_BUTTON.filter}
           placeHolder={FILTER_BUTTON.placeHolder}
           choices={FILTER_BUTTON.choices}
-          handleFilters={handleFilters}
+          filterHandler={filterHandler}
+          movePageHandler={movePageHandler}
         />
       ))}
       <div
-        onClick={() => handleFilters('all', 'reset')}
+        onClick={() => filterHandler('all', 'reset')}
         className='flex cursor-pointer items-center gap-2 text-tbGray hover:text-tbRed'
       >
         <LucideIcon name='RotateCw' size={20} />
